Update readme, add possibility to run shell commands on set and get active config

+++ backend/filemanagement.py
@@ -77,21 +77,36 @@ def get_yaml_as_json(request, path):
     return validator.get_config()
 
 
-def get_active_config(active_config):
-    if islink(active_config) and isfile(active_config):
-        target = os.readlink(active_config)
-        head, tail = split(target)
-        return tail
+def get_active_config(request):
+    active_config = request.app["active_config"]
+    on_get = request.app["on_get_active_config"]
+    if not on_get:
+        if islink(active_config) and isfile(active_config):
+            target = os.readlink(active_config)
+            _head, tail = split(target)
+            return tail
+        else:
+            return None
     else:
-        return None
+        stream = os.popen(on_get)
+        target = stream.read().strip()
+        _head, tail = split(target)
+        return tail
 
 
-def set_as_active_config(active_config, file):
-    if not active_config:
-        return
-    if islink(active_config):
-        os.unlink(active_config)
-    os.symlink(file, active_config)
+
+def set_as_active_config(request, file):
+    active_config = request.app["active_config"]
+    update = request.app["update_symlink"]
+    on_set = request.app["on_set_active_config"]
+    if update:
+        if not active_config:
+            return
+        if islink(active_config):
+            os.unlink(active_config)
+        os.symlink(file, active_config)
+    if on_set:
+        os.system(f"{on_set} {file}")
 
 
 def save_config(config_name, json_config, request):
